9-07-26_mutasiantarpt

+++ app/Http/Controllers/AccessController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Models\Access;
+use App\Models\Perusahaan;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\Rule;
@@ -15,35 +16,62 @@ class AccessController extends Controller
    */
   public function index(Request $request)
   {
-    $search = $request->search;
-    $kategori = $request->kategori;
-    $jenis = $request->jenis;
-    $status = $request->status;
+    $user = auth()->user();
 
-    $query = Access::query();
+    $query = Access::with('perusahaan');
 
-    if ($search) {
-      $query->where('nama_akses', 'like', "%{$search}%");
+    /*
+    |--------------------------------------------------------------------------
+    | SUPER ADMIN
+    |--------------------------------------------------------------------------
+    */
+
+    if ($user->role == 'super_admin') {
+      if ($request->filled('perusahaan')) {
+        $query->where('id_perusahaan', $request->perusahaan);
+      }
+    } /*
+    |--------------------------------------------------------------------------
+    | PETUGAS
+    |--------------------------------------------------------------------------
+    */ else {
+      $query->where('id_perusahaan', $user->id_perusahaan);
     }
 
-    if ($kategori) {
-      $query->where('kategori', $kategori);
+    /*
+    |--------------------------------------------------------------------------
+    | FILTER
+    |--------------------------------------------------------------------------
+    */
+
+    if ($request->filled('jenis')) {
+      $query->where('jenis', $request->jenis);
     }
 
-    if ($jenis) {
-      $query->where('jenis', $jenis);
+    if ($request->filled('kategori')) {
+      $query->where('kategori', $request->kategori);
     }
 
-    if ($status) {
-      $query->where('status', $status);
+    if ($request->filled('status')) {
+      $query->where('status', $request->status);
+    }
+
+    if ($request->filled('search')) {
+      $search = $request->search;
+
+      $query->where(function ($q) use ($search) {
+        $q->where('nama_akses', 'like', "%{$search}%")->orWhere('keterangan', 'like', "%{$search}%");
+      });
     }
 
     $accesses = $query
-      ->latest()
-      ->paginate(10)
-      ->appends($request->except('page'));
+      ->orderBy('nama_akses')
+      ->paginate(15, ['*'], 'page', request()->input('page', 1))
+      ->appends(request()->query());
 
-    return view('content.dashboard.hak_akses.index', compact('accesses'));
+    $perusahaans = Perusahaan::orderBy('nama_perusahaan')->get();
+
+    return view('content.dashboard.hak_akses.index', compact('accesses', 'perusahaans'));
   }
 
   /**
@@ -51,22 +79,27 @@ class AccessController extends Controller
    */
   public function store(Request $request)
   {
-    $request->validate([
+    $user = auth()->user();
+
+    $rules = [
       'kategori' => 'required|in:Aplikasi,Hak Akses',
       'jenis' => 'required|in:Software,PPN,NON PPN',
-      'nama_akses' => [
-        'required',
-        'string',
-        'max:150',
-
-        Rule::unique('accesses')->where(function ($query) use ($request) {
-          return $query->where('kategori', $request->kategori)->where('jenis', $request->jenis);
-        }),
-      ],
       'status' => 'required|in:aktif,nonaktif',
-    ]);
+      'nama_akses' => ['required', 'string', 'max:150'],
+    ];
 
-    // Validasi kombinasi kategori & jenis
+    if ($user->role == 'super_admin') {
+      $rules['id_perusahaan'] = 'required|exists:perusahaans,id';
+    }
+
+    $request->validate($rules);
+
+    /*
+    |--------------------------------------------------------------------------
+    | VALIDASI KATEGORI
+    |--------------------------------------------------------------------------
+    */
+
     if ($request->kategori == 'Aplikasi' && $request->jenis != 'Software') {
       return back()
         ->withErrors([
@@ -83,8 +116,37 @@ class AccessController extends Controller
         ->withInput();
     }
 
+    /*
+    |--------------------------------------------------------------------------
+    | PERUSAHAAN
+    |--------------------------------------------------------------------------
+    */
+
+    $idPerusahaan = $user->role == 'super_admin' ? $request->id_perusahaan : $user->id_perusahaan;
+
+    /*
+    |--------------------------------------------------------------------------
+    | DUPLIKAT
+    |--------------------------------------------------------------------------
+    */
+
+    $cek = Access::where('id_perusahaan', $idPerusahaan)
+      ->where('kategori', $request->kategori)
+      ->where('jenis', $request->jenis)
+      ->where('nama_akses', strtoupper($request->nama_akses))
+      ->exists();
+
+    if ($cek) {
+      return back()
+        ->withInput()
+        ->withErrors([
+          'nama_akses' => 'Hak Akses sudah ada pada perusahaan tersebut.',
+        ]);
+    }
+
     try {
       Access::create([
+        'id_perusahaan' => $idPerusahaan,
         'kategori' => $request->kategori,
         'jenis' => $request->jenis,
         'nama_akses' => strtoupper($request->nama_akses),
@@ -133,24 +195,22 @@ class AccessController extends Controller
   public function update(Request $request, string $id)
   {
     $access = Access::findOrFail($id);
+    $user = auth()->user();
+    if ($user->role == 'super_admin') {
+      $request->validate([
+        'id_perusahaan' => 'required|exists:perusahaans,id',
+      ]);
+    }
+
+    if ($user->role != 'super_admin' && $access->id_perusahaan != $user->id_perusahaan) {
+      abort(403);
+    }
 
     $request->validate([
-      'kategori' => 'required|in:Aplikasi,Hak Akses',
-      'jenis' => 'required|in:Software,PPN,NON PPN',
-      'nama_akses' => [
-        'required',
-
-        'string',
-
-        'max:150',
-
-        Rule::unique('accesses')
-          ->ignore($id)
-          ->where(function ($query) use ($request) {
-            return $query->where('kategori', $request->kategori)->where('jenis', $request->jenis);
-          }),
-      ],
-      'status' => 'required|in:aktif,nonaktif',
+      'kategori' => ['required', 'in:Aplikasi,Hak Akses'],
+      'jenis' => ['required', 'in:Software,PPN,NON PPN'],
+      'nama_akses' => ['required', 'string', 'max:150'],
+      'status' => ['required', 'in:aktif,nonaktif'],
     ]);
 
     // Validasi kombinasi kategori & jenis
@@ -169,9 +229,26 @@ class AccessController extends Controller
         ])
         ->withInput();
     }
+    $idPerusahaan = $user->role == 'super_admin' ? $request->id_perusahaan : $user->id_perusahaan;
+
+    $cek = Access::where('id_perusahaan', $idPerusahaan)
+      ->where('kategori', $request->kategori)
+      ->where('jenis', $request->jenis)
+      ->where('nama_akses', strtoupper($request->nama_akses))
+      ->where('id', '!=', $access->id)
+      ->exists();
+
+    if ($cek) {
+      return back()
+        ->withInput()
+        ->withErrors([
+          'nama_akses' => 'Hak Akses sudah ada pada perusahaan tersebut.',
+        ]);
+    }
 
     try {
       $access->update([
+        'id_perusahaan' => $idPerusahaan,
         'kategori' => $request->kategori,
         'jenis' => $request->jenis,
         'nama_akses' => strtoupper($request->nama_akses),
@@ -194,6 +271,12 @@ class AccessController extends Controller
     try {
       $access = Access::findOrFail($id);
 
+      $user = auth()->user();
+
+      if ($user->role != 'super_admin' && $access->id_perusahaan != $user->id_perusahaan) {
+        abort(403);
+      }
+
       $access->delete();
 
       return back()->with('success', 'Hak Akses & Aplikasi berhasil dihapus.');
@@ -201,6 +284,30 @@ class AccessController extends Controller
       Log::error($e->getMessage());
 
       return back()->with('error', 'Gagal menghapus data.');
+    }
+  }
+  public function filterJenis(Request $request, $jenis)
+  {
+    try {
+      $user = auth()->user();
+
+      $query = Access::where('jenis', $jenis)->where('status', 'aktif');
+
+      if ($user->role == 'super_admin') {
+        if ($request->filled('perusahaan')) {
+          $query->where('id_perusahaan', $request->perusahaan);
+        } else {
+          return response()->json([]);
+        }
+      } else {
+        $query->where('id_perusahaan', $user->id_perusahaan);
+      }
+
+      return response()->json($query->orderBy('nama_akses')->get(['id', 'nama_akses']));
+    } catch (\Exception $e) {
+      Log::error($e->getMessage());
+
+      return response()->json([], 500);
     }
   }
 }
